Remove non-interface properties from main library
Don't add modules or their compile options as interface compile options
or interface link libraries to the main library. The latter doesn't make
sense, since object libraries aren't part of the link interface anyway,
and the former causes the module targets to become interface
dependencies of the main library (which, again, is a problem because
trying to export object libraries doesn't make sense). The only apparent
use case of transitive compile options is anyway for coverage-related
stuff, and consumers probably don't want to pick that up from using
Valhalla (or even use a version of the library built with coverage).

This creates a minor issue, however, as some components have additional
dependencies that are public. To address this, create a new interface
library, valhalla-data-services, which conveys the necessary link
dependencies. This is being added separately as not all users will need
these, and not all configurations of valhalla have such dependencies.

In addition, the services need to link to libprime_server. Previously,
this was "leaked" as a public dependency, even though most library users
don't need it. Fix this by modifying the services to additionally link
directly to libprime_server.