Start implementing exported targets
Start adding CMake exported targets. This isn't functional, yet, due to
issues with third-party dependencies, but it's a step to making it much,
much easier for other CMake-based consumers to use Valhalla. This will
also make it possible for such consumers to use a build tree, without
having to perform an installation.

In addition to generating the required files, fix the interface includes
so that build-tree includes are not "leaked" to the install interface
(which, besides being incorrect, makes CMake grumpy).