Hide robin_hood from the main library
Tweak how we bring in robin_hood as a dependency to avoid it being
visible as a link dependency of the main library. We can do this because
we only need it when building the object code of the object library
components that get assembled into the main library. This is necessary
so that it doesn't leak into the main library's link interface, since we
don't export it; otherwise, CMake will complain about a missing link
dependency when we go to export the main library. (This happens when
building as a static library rather than a shared library, since static
libraries aren't actually linked and therefore have no such thing as
"private" dependencies; ALL of their link dependencies are effectively
public.)